tests: clean up pmac tests
