+++ tests/epics/pmac/test_pmac_trajectory_generation.py
@@ -3,7 +3,6 @@ import pytest
 from scanspec.core import Path
 from scanspec.specs import Fly, Line, Spiral
 
-from ophyd_async.core import init_devices
 from ophyd_async.epics.motor import Motor
 from ophyd_async.epics.pmac import (
     PmacIO,
@@ -15,47 +14,26 @@ from ophyd_async.epics.pmac._pmac_trajectory_generation import (
 from ophyd_async.epics.pmac._utils import (
     _PmacMotorInfo,  # noqa: PLC2701
 )
-from ophyd_async.testing import set_mock_value
 
 
 @pytest.fixture
-async def sim_motors():
-    async with init_devices(mock=True):
-        sim_x_motor = Motor("BLxxI-MO-STAGE-01:X")
-        sim_y_motor = Motor("BLxxI-MO-STAGE-01:Y")
-        sim_pmac = PmacIO(
-            prefix="Test_PMAC",
-            raw_motors=[sim_x_motor, sim_y_motor],
-            coord_nums=[],
-        )
-
-    pmac_x = sim_pmac.assignment[sim_pmac.motor_assignment_index[sim_x_motor]]
-    pmac_y = sim_pmac.assignment[sim_pmac.motor_assignment_index[sim_y_motor]]
-    set_mock_value(pmac_x.cs_port, "CS1")
-    set_mock_value(pmac_x.cs_number, 1)
-    set_mock_value(pmac_x.cs_axis_letter, "X")
-    set_mock_value(pmac_y.cs_port, "CS1")
-    set_mock_value(pmac_y.cs_number, 1)
-    set_mock_value(pmac_y.cs_axis_letter, "Y")
-    set_mock_value(sim_x_motor.acceleration_time, 0.5)
-    set_mock_value(sim_x_motor.max_velocity, 5)
-    set_mock_value(sim_y_motor.acceleration_time, 0.5)
-    set_mock_value(sim_y_motor.max_velocity, 10)
-
-    yield (sim_pmac, sim_x_motor, sim_y_motor)
+async def motor_info(sim_motors: tuple[PmacIO, Motor, Motor]):
+    _, sim_x_motor, sim_y_motor = sim_motors
+    yield _PmacMotorInfo(
+        "CS1",
+        1,
+        {sim_x_motor: 6, sim_y_motor: 7},
+        {sim_x_motor: 10, sim_y_motor: 10},
+        {sim_x_motor: 5, sim_y_motor: 5},
+    )
 
 
-async def test_line_trajectory_from_slice(sim_motors: tuple[PmacIO, Motor, Motor]):
+async def test_line_trajectory_from_slice(
+    sim_motors: tuple[PmacIO, Motor, Motor], motor_info: _PmacMotorInfo
+):
     _, sim_x_motor, _ = sim_motors
     spec = Fly(2.0 @ Line(sim_x_motor, 1, 5, 9))
     slice = Path(spec.calculate()).consume()
-    motor_info = _PmacMotorInfo(
-        "CS1",
-        1,
-        {sim_x_motor: 6},
-        {sim_x_motor: 10},
-        {sim_x_motor: 5},
-    )
     trajectory, exit_pvt = Trajectory.from_slice(slice, motor_info, ramp_up_time=2)
     trajectory.append_ramp_down(exit_pvt, {sim_x_motor: np.float64(6)}, 2, 0)
 
@@ -161,17 +139,12 @@ async def test_line_trajectory_from_slice(sim_motors: tuple[PmacIO, Motor, Motor
     )
 
 
-async def test_spiral_trajectory_from_slice(sim_motors: tuple[PmacIO, Motor, Motor]):
+async def test_spiral_trajectory_from_slice(
+    sim_motors: tuple[PmacIO, Motor, Motor], motor_info: _PmacMotorInfo
+):
     _, sim_x_motor, sim_y_motor = sim_motors
     spec = Spiral(sim_x_motor, sim_y_motor, 0, 0, 5, 5, 3)
     slice = Path(Fly(2.0 @ spec).calculate()).consume()
-    motor_info = _PmacMotorInfo(
-        "CS1",
-        1,
-        {sim_x_motor: 6, sim_y_motor: 7},
-        {sim_x_motor: 10, sim_y_motor: 10},
-        {sim_x_motor: 5, sim_y_motor: 5},
-    )
 
     trajectory, exit_pvt = Trajectory.from_slice(slice, motor_info, ramp_up_time=2.0)
     trajectory.append_ramp_down(
@@ -233,26 +206,20 @@ async def test_spiral_trajectory_from_slice(sim_motors: tuple[PmacIO, Motor, Mot
     }
 
     assert trajectory.durations == pytest.approx(
-        [2.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 2.0], 1e-5
+        [2.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 2.0]
     )
 
-    assert trajectory.user_programs == pytest.approx([1, 1, 1, 1, 1, 1, 1, 8])
+    assert (trajectory.user_programs == [1, 1, 1, 1, 1, 1, 1, 8]).all()
 
 
-async def test_snaked_trajectory_with_gaps(sim_motors: tuple[PmacIO, Motor, Motor]):
+async def test_snaked_trajectory_with_gaps(
+    sim_motors: tuple[PmacIO, Motor, Motor], motor_info: _PmacMotorInfo
+):
     _, sim_x_motor, sim_y_motor = sim_motors
     spec = Fly(1.0 @ (Line(sim_y_motor, 10, 12, 3) * ~Line(sim_x_motor, 1, 5, 5)))
     slice = Path(spec.calculate()).consume()
-    motor_info = _PmacMotorInfo(
-        "CS1",
-        1,
-        {sim_x_motor: 6, sim_y_motor: 7},
-        {sim_x_motor: 10, sim_y_motor: 10},
-        {sim_x_motor: 5, sim_y_motor: 5},
-    )
 
     trajectory, exit_pvt = Trajectory.from_slice(slice, motor_info, ramp_up_time=1.0)
-
     trajectory.append_ramp_down(
         exit_pvt, {sim_x_motor: np.float64(6.0), sim_y_motor: np.float64(12)}, 1.0, 0
     )
@@ -437,8 +404,9 @@ async def test_snaked_trajectory_with_gaps(sim_motors: tuple[PmacIO, Motor, Moto
         ]
     )
 
-    assert trajectory.user_programs == pytest.approx(
-        [
+    assert (
+        trajectory.user_programs
+        == [
             1,
             1,
             1,
@@ -480,69 +448,63 @@ async def test_snaked_trajectory_with_gaps(sim_motors: tuple[PmacIO, Motor, Moto
             1,
             8,
         ]
-    )
+    ).all()
 
     assert trajectory.durations == pytest.approx(
         [
-            1.000000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.100000,
+            1.0,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.1,
             0.216227,
             0.216227,
-            0.100000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.100000,
+            0.1,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.1,
             0.216227,
             0.216227,
-            0.100000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            1.000000,
+            0.1,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            1.0,
         ],
         1e-5,
     )
 
 
-async def test_grid_trajectory_with_gaps(sim_motors: tuple[PmacIO, Motor, Motor]):
+async def test_grid_trajectory_with_gaps(
+    sim_motors: tuple[PmacIO, Motor, Motor], motor_info: _PmacMotorInfo
+):
     _, sim_x_motor, sim_y_motor = sim_motors
     spec = Fly(1.0 @ (Line(sim_y_motor, 10, 12, 3) * Line(sim_x_motor, 1, 5, 5)))
     slice = Path(spec.calculate()).consume()
-    motor_info = _PmacMotorInfo(
-        "CS1",
-        1,
-        {sim_x_motor: 6, sim_y_motor: 7},
-        {sim_x_motor: 10, sim_y_motor: 10},
-        {sim_x_motor: 5, sim_y_motor: 5},
-    )
 
     trajectory, exit_pvt = Trajectory.from_slice(slice, motor_info, ramp_up_time=1.0)
-
     trajectory.append_ramp_down(
         exit_pvt,
         {sim_x_motor: np.float64(6.0), sim_y_motor: np.float64(12.0)},
@@ -738,8 +700,9 @@ async def test_grid_trajectory_with_gaps(sim_motors: tuple[PmacIO, Motor, Motor]
         ]
     )
 
-    assert trajectory.user_programs == pytest.approx(
-        [
+    assert (
+        trajectory.user_programs
+        == [
             1,
             1,
             1,
@@ -783,68 +746,63 @@ async def test_grid_trajectory_with_gaps(sim_motors: tuple[PmacIO, Motor, Motor]
             1,
             8,
         ]
-    )
+    ).all()
 
     assert trajectory.durations == pytest.approx(
         [
-            1.000000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
+            1.0,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
             0.060250,
             0.539749,
             0.520000,
             0.539749,
             0.060250,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
             0.060250,
             0.539749,
             0.520000,
             0.539749,
             0.060250,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            0.500000,
-            1.000000,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            0.5,
+            1.0,
         ],
         1e-5,
     )
 
 
-async def test_from_gap(sim_motors):
+async def test_from_gap(
+    sim_motors: tuple[PmacIO, Motor, Motor], motor_info: _PmacMotorInfo
+):
     _, sim_x_motor, sim_y_motor = sim_motors
     spec = Fly(1.0 @ (Line(sim_y_motor, 10, 12, 3) * ~Line(sim_x_motor, 1, 5, 5)))
     slice = Path(spec.calculate()).consume()
-    motor_info = _PmacMotorInfo(
-        "CS1",
-        1,
-        {sim_x_motor: 6, sim_y_motor: 7},
-        {sim_x_motor: 10, sim_y_motor: 10},
-        {sim_x_motor: 5, sim_y_motor: 5},
-    )
 
     # This is info about the frame just before a gap is created.
     entry_pvt = PVT(
@@ -866,7 +824,6 @@ async def test_from_gap(sim_motors):
             5.5,
             5.0,
         ],
-        1e-5,
     )
 
     assert trajectory.positions[sim_y_motor] == pytest.approx(
@@ -878,7 +835,6 @@ async def test_from_gap(sim_motors):
             11.0,
             11.0,
         ],
-        1e-5,
     )
 
     assert trajectory.velocities[sim_x_motor] == pytest.approx(
@@ -890,7 +846,6 @@ async def test_from_gap(sim_motors):
             -1.0,
             -1.0,
         ],
-        1e-5,
     )
 
     assert trajectory.velocities[sim_y_motor] == pytest.approx(
@@ -902,25 +857,24 @@ async def test_from_gap(sim_motors):
             0.0,
             0.0,
         ],
-        1e-5,
     )
 
     assert trajectory.durations == pytest.approx(
         [
-            1.000000,
-            0.100000,
+            1.0,
+            0.1,
             0.216227,
             0.216227,
-            0.100000,
-            0.500000,
+            0.1,
+            0.5,
         ],
         1e-5,
     )
 
-    assert trajectory.user_programs == pytest.approx([1, 2, 2, 2, 1, 1])
+    assert (trajectory.user_programs == [1, 2, 2, 2, 1, 1]).all()
 
 
-async def test_from_collection_window(sim_motors):  # noqa: D103
+async def test_from_collection_window(sim_motors: tuple[PmacIO, Motor, Motor]):
     _, _, sim_y_motor = sim_motors
     spec = Fly(1.0 @ (Line(sim_y_motor, 1, 5, 5)))
     path = Path(spec.calculate())
@@ -938,37 +892,26 @@ async def test_from_collection_window(sim_motors):  # noqa: D103
 
     assert trajectory.positions[sim_y_motor] == pytest.approx(
         [1.5, 2, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0],
-        1e-5,
     )
 
     assert trajectory.velocities[sim_y_motor] == pytest.approx(
         [0.75, 1, 1, 1, 1, 1, 1, 1],
-        1e-5,
     )
 
     assert trajectory.durations == pytest.approx(
         [0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5],
-        1e-5,
     )
 
-    assert trajectory.user_programs == pytest.approx(
-        [1, 1, 1, 1, 1, 1, 1, 1],
-        1e-5,
-    )
+    assert (trajectory.user_programs == [1, 1, 1, 1, 1, 1, 1, 1]).all()
 
 
-async def test_appending_trajectory(sim_motors):  # noqa: D103
+async def test_appending_trajectory(
+    sim_motors: tuple[PmacIO, Motor, Motor], motor_info: _PmacMotorInfo
+):
     _, sim_x_motor, sim_y_motor = sim_motors
     spec = Fly(1.0 @ (Line(sim_y_motor, 10, 11, 2) * ~Line(sim_x_motor, 1, 3, 3)))
     path = Path(spec.calculate())
     slice = path.consume(3)
-    motor_info = _PmacMotorInfo(
-        "CS1",
-        1,
-        {sim_x_motor: 6, sim_y_motor: 7},
-        {sim_x_motor: 10, sim_y_motor: 10},
-        {sim_x_motor: 5, sim_y_motor: 5},
-    )
 
     first_trajectory, first_exit_pvt = Trajectory.from_slice(
         slice, motor_info, ramp_up_time=1
@@ -1080,7 +1023,6 @@ async def test_appending_trajectory(sim_motors):  # noqa: D103
             0.0,
             0.0,
         ],
-        1e-5,
     )
 
     assert overall_trajectory.durations == pytest.approx(
@@ -1107,8 +1049,9 @@ async def test_appending_trajectory(sim_motors):  # noqa: D103
         1e-5,
     )
 
-    assert overall_trajectory.user_programs == pytest.approx(
-        [
+    assert (
+        overall_trajectory.user_programs
+        == [
             1,
             1,
             1,
@@ -1127,23 +1070,18 @@ async def test_appending_trajectory(sim_motors):  # noqa: D103
             1,
             1,
             8,
-        ],
-        1e-5,
-    )
+        ]
+    ).all()
 
 
-async def test_hardware_triggered_step_scan(sim_motors):
+async def test_hardware_triggered_step_scan(
+    sim_motors: tuple[PmacIO, Motor, Motor], motor_info: _PmacMotorInfo
+):
     _, sim_x_motor, _ = sim_motors
     spec = 1.0 @ Line(sim_x_motor, 1, 3, 3)
     path = Path(spec.calculate())
     slice = path.consume()
-    motor_info = _PmacMotorInfo(
-        "CS1",
-        1,
-        {sim_x_motor: 6},
-        {sim_x_motor: 10},
-        {sim_x_motor: 5},
-    )
+
     trajectory, exit_pvt = Trajectory.from_slice(slice, motor_info, ramp_up_time=2)
     trajectory.append_ramp_down(exit_pvt, {sim_x_motor: np.float64(6)}, 2, 0)
